refactor: refactor route
