+++ src/route/index.ts
@@ -1,10 +1,11 @@
 import { Hono } from "hono";
 import { serveStatic } from "hono/deno";
 
-import { home, redirect } from "../controllers/index.ts";
-import { sendAccessLog } from "../services/log.ts";
+import { Logger } from "../service/logger.ts";
+import { home, redirect } from "../controller/index.ts";
 
 const app = new Hono();
+const logger = new Logger();
 
 app.use(async (c, next) => {
   await next();
@@ -12,9 +13,9 @@ app.use(async (c, next) => {
   if (!c.req.path.startsWith("/api")) {
     // 200-299 or 300-399
     if (c.res.ok || (c.res.status >= 300 && c.res.status < 400)) {
-      await sendAccessLog(c, false);
+      await logger.access(c, false);
     } else {
-      await sendAccessLog(c, true);
+      await logger.access(c, true);
     }
   }
 });
